requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.4
plotly==5.18.0
Pillow==10.2.0 